add search by stores
